+++ include/simol/statphys/system/Chain.hpp
@@ -20,8 +20,8 @@ namespace simol
     public:
       BiChain(Input const& input);
       void computeAllForces();
-      virtual void computeProfile(Output& output, Dynamics const& dyna, int iOfStep) const;
-      virtual void computeProfile(Output& output, LangevinBase const& model, int iOfStep) const;
+      virtual void computeProfile(Output& output, Dynamics const& dyna, long int iOfStep) const;
+      virtual void computeProfile(Output& output, LangevinBase const& model, long int iOfStep) const;
     protected:
       Particle ancorParticle_;
   };
@@ -34,8 +34,8 @@ namespace simol
       void triInteraction(Particle& particle1, Particle& particle2, Particle& particle3) const;
       void computeAllForces();
       virtual double boundaryPotEnergy() const;
-      virtual void computeProfile(Output& output, Dynamics const& dyna, int iOfStep) const;
-      virtual void computeProfile(Output& output, LangevinBase const& model, int iOfStep) const;
+      virtual void computeProfile(Output& output, Dynamics const& dyna, long int iOfStep) const;
+      virtual void computeProfile(Output& output, LangevinBase const& model, long int iOfStep) const;
     protected:
       Particle ancorParticle1_;
       Particle ancorParticle2_;
