+++ include/simol/statphys/system/System.hpp
@@ -59,7 +59,7 @@ namespace simol
       virtual double computeMeanPotLaw(double betaLocal) const;
 
       //-- output functions --
-      virtual void computeProfile(Output& /*output*/, Dynamics const& /*model*/, int /*iOfStep*/)const;
+      virtual void computeProfile(Output& /*output*/, Dynamics const& /*model*/, long int /*iOfStep*/)const;
 
       // currently specific to chains
       virtual double boundaryPotEnergy() const;
